parkhang/requirements/base.txt
# Requirements for all environments

# General
requests==2.27.1
whitenoise==5.3.0  # https://github.com/evansd/whitenoise
redis==4.2.2  # https://github.com/redis/redis-py


# Django
# -----------------------------------------------------------------------
Django==3.2.12
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-webpack-loader==1.4.1
django-redis==5.2.0  # https://github.com/jazzband/django-redis
django-extensions==3.1.5  # https://github.com/django-extensions/django-extensions
# django-cors-headers==3.4.0  # https://github.com/adamchainz/django-cors-headers

# Django REST Framework
djangorestframework==3.13.1
